+++ core-backup/src/main/scala/io/aiven/guardian/kafka/backup/configs/Backup.scala
@@ -8,8 +8,11 @@ import scala.concurrent.duration.FiniteDuration
   *   Determines how the backed up objects/files are segregated depending on a time configuration
   * @param commitTimeoutBufferWindow
   *   A buffer that is added ontop of the `timeConfiguration` when setting the Kafka Consumer commit timeout.
+  * @param compression
+  *   Which compression to use for the backed up data
   */
 final case class Backup(kafkaGroupId: String,
                         timeConfiguration: TimeConfiguration,
-                        commitTimeoutBufferWindow: FiniteDuration
+                        commitTimeoutBufferWindow: FiniteDuration,
+                        compression: Option[Compression]
 )
