+++ core-backup/src/main/scala/io/aiven/guardian/kafka/backup/configs/Compression.scala
@@ -1,0 +1,5 @@
+package io.aiven.guardian.kafka.backup.configs
+
+import io.aiven.guardian.kafka.models.CompressionType
+
+final case class Compression(`type`: CompressionType, level: Option[Int])
