+++ core-backup/src/test/scala/io/aiven/guardian/kafka/backup/GzipCompressionBackupClientInterfaceSpec.scala
@@ -1,0 +1,12 @@
+package io.aiven.guardian.kafka.backup
+
+import akka.actor.ActorSystem
+import io.aiven.guardian.akka.AnyPropTestKit
+import io.aiven.guardian.kafka.backup.configs.Compression
+import io.aiven.guardian.kafka.models.Gzip
+
+class GzipCompressionBackupClientInterfaceSpec
+    extends AnyPropTestKit(ActorSystem("GzipCompressionBackupClientInterfaceSpec"))
+    with BackupClientInterfaceTest {
+  override val compression: Option[Compression] = Some(Compression(Gzip, None))
+}
