+++ core/src/main/scala/io/aiven/guardian/kafka/models/BackupObjectMetadata.scala
@@ -1,0 +1,11 @@
+package io.aiven.guardian.kafka.models
+
+final case class BackupObjectMetadata(compression: Option[CompressionType])
+
+object BackupObjectMetadata {
+  def fromKey(key: String): BackupObjectMetadata =
+    if (key.endsWith(".gz"))
+      BackupObjectMetadata(Some(Gzip))
+    else
+      BackupObjectMetadata(None)
+}
