+++ core/src/main/scala/io/aiven/guardian/kafka/models/CompressionType.scala
@@ -1,0 +1,9 @@
+package io.aiven.guardian.kafka.models
+
+sealed trait CompressionType {
+  val pretty: String
+}
+
+case object Gzip extends CompressionType {
+  override val pretty: String = "Gzip"
+}
